update code into more granular subgroups. also moved nrt update specific group code to its own concern.
